I HAVE CHANGED PIECES - READ THIS
K so camilo's recursion is either broken or unfathomably slow even at
depth = 1. I am currently looking into why.

Emerson, I made a change in pieces.py: lines 111 and ~142, you append
the flying general move to successors, but you appended it as a one
element list of a tuple [(x,y)] when your function requires a tuple.
This led to check() comparing tuple to int for that one case and caused
it to error out. I changed it to just append the tuple. It runs now and
seems to work.

PLEASE MAKE SURE THIS IS FINE.

And camilo, as soon as you get the chance, this recursion looks like it
needs debugging and you know it better than I

Sorry and thanks