Added optional eliminated parameter to Gamestate
Eliminated is an optional parameter defaulting to None if not provided. self.eliminated holds the piece that got eliminated in creating this gamestate. Adam, please modify your code in game.py for parts where you eliminate a piece. (Note because it defaults to None, you don't need to change every call)